refactored specializations module 2

+++ src/specialization/specialization.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateSpecializationDto } from './dto/create-specialization.dto';
-import { UpdateSpecializationDto } from './dto/update-specialization.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { UpdateSpecializationDto } from './dto/update-specialization.dto';
 
 @Injectable()
 export class SpecializationService {
@@ -14,33 +14,65 @@ export class SpecializationService {
     return specialization;
   }
 
-  findAllByTitle(title: string) {
-    if (!title) {
-      return this.prisma.specialization.findMany();
-    }
-    return this.prisma.specialization.findMany({
-      where: {
-        title: {
-          contains: title,
-        },
-      },
+  async findAllByTitle(title: string, page = 1, take = 10) {
+    const conditions = title
+      ? {
+          where: {
+            title: {
+              contains: title,
+              mode: 'insensitive',
+            },
+          },
+        }
+      : null;
+    const specializations = await this.prisma.specialization.findMany({
+      ...conditions,
+      skip: (page - 1) * take,
+      take,
     });
+
+    const totalCount = await this.prisma.specialization.count({
+      ...conditions,
+    });
+
+    return {
+      data: specializations,
+      pagination: {
+        page,
+        take,
+        total: totalCount,
+      },
+    };
+  }
+
+  private async isSpecializationExist(id: string) {
+    const specialization = await this.prisma.specialization.findUnique({
+      where: { id },
+    });
+    if (!specialization) {
+      throw new NotFoundException(`Specialization with ID ${id} not found`);
+    }
+    return true;
   }
 
   update(id: string, updateSpecializationDto: UpdateSpecializationDto) {
-    return this.prisma.specialization.update({
-      where: {
-        id,
-      },
-      data: updateSpecializationDto,
-    });
+    if (this.isSpecializationExist(id)) {
+      return this.prisma.specialization.update({
+        where: {
+          id,
+        },
+        data: updateSpecializationDto,
+      });
+    }
   }
 
   remove(id: string) {
-    return this.prisma.specialization.delete({
-      where: {
-        id,
-      },
-    });
+    if (this.isSpecializationExist(id)) {
+      return this.prisma.specialization.delete({
+        where: {
+          id,
+        },
+      });
+    }
   }
 }
